[#906] Separate upgrade modal UI from Domains screen
